Add files via upload
Historically, it has been difficult to make data on human trafficking readily accessible to analysts, academics, practitioners and policy-makers. Data on human trafficking are often highly sensitive matter and at the same time it is widely recognized that one of the foremost challenges in developing targeted counter-trafficking responses and measuring their impact is the lack of reliable, high-quality information.

To that end, the Counter-Trafficking Data Collaborative (CTDC) is the first global data hub on human trafficking, publishing harmonized data from counter-trafficking organizations around the world in order to equip the counter-trafficking community with up to date, reliable data on human trafficking.

This dataset contains information on 48.8k victims of human trafficking, including the reason, means of control, origin and destination, as well as other variables. Note that missingness is displayed as -99 in this dataset.