Add files via upload
Female State Legislators
Description: number of female state legislators from all 50 states for 1981-2019

Source: Rutgers

Processing: data before 1981 was removed from the original .csv because it was not complete, some columns names were changed and all "-" were changed to a 0, and the formatting for Nebraska was changed

Columns:

State (string): name of each state
Asterisk for Nebraska throughout the dataset: Nebraska has a unicameral legislature, meaning they only have one chamber, not the traditional two chambers that all other states use; house (sometimes referred to as an “assembly”) and senate. Nebraska refers to its members as Senators, so the data is reflected in the Senate columns and the House columns are all listed with zeros.They are also nonpartisan, so their party affiliation is not reflected in their roles as legislators, however most members are registered with a major political party. Since members are nonpartisan the party affiliations for all cells related to Nebraska are marked as zero.
Year (numerical): year of the legislative session, from 1981 and 2019.
National Rank (numerical): national rank that year on a scale from 1-50, the state ranked 1 has the highest percent of female legislators and the state ranked 50 has the lowest percent of female legislators
Female Senate Democrats (numerical): number of female Democrats in the State Senate in that given state that given year.
Female Senate Republicans (numerical): number of female Republicans in the State Senate in that given state that given year.
Female Senate Independents (numerical): number of female Independents in the State Senate in that given state that given year.
Female Senate Other (numerical): number of female State Senators registered with a party other than Democrat, Republican, and Independent.
Female Senate Total (numerical): total number of female state senators in that state for that given year.
Total Senate Members(numerical): total number of senators in the state senate for that state for that given year.
Female House Democrats (numerical): number of female Democrats in the State House in that given state that given year.
Female House Republicans (numerical): number of female Republicans in the State House of Representatives in that given state that given year.
Female House Independents (numerical): number of female Independents in the State House of Representatives in that state in that particular year.
Female House Other (numerical): number of Female State Representatives in the State House of Representatives registered with a party other than Democrat, Republican, an Independent.
Female House Total (number): total number of female state house representatives in that state for that given year.
Total House Members (numerical): total number of representatives in the state house for that state for that given year.
Total Females in Legislature (numerical): total number of house and senate female members in that state in that given year.
Total Legislature Members (numerical): total number of legislators (combining house and senate).
Percentage of Female in Legislature (numerical): total number of female legislators divided by the total legislature.