Add files via upload
International Exhibition of Modern Art 1913
Description: Data on the works of art included in the International Exhibition of Modern Art 1913

The Armory Show, also known as the International Exhibition of Modern Art, was a show organized by the Association of American Painters and Sculptors in 1913. It was the first large exhibition of modern art in America...

The three-city exhibition started in New York City's 69th Regiment Armory, on Lexington Avenue between 25th and 26th Streets, from February 17 until March 15, 1913. The exhibition went on to the Art Institute of Chicago and then to The Copley Society of Art in Boston, where, due to a lack of space, all the work by American artists was removed.

The show became an important event in the history of American art, introducing astonished Americans, who were accustomed to realistic art, to the experimental styles of the European avant garde, including Fauvism, Cubism, and Futurism. The show served as a catalyst for American artists, who became more independent and created their own "artistic language." Source accessed January 12, 2020

Data Source: Dr. Nancy Ross and Her Students via Bootstrap

Columns:

Title (string): title of the work of art
Artist (string): name of the artist
Sex (string): sex of the artist
Medium (string): the medium the work of art was done in
Location (string): the names of the cities where the work of art was exhibited
Number of Locations (numerical): the number locations where the work of art was exhibited
Sale Price in Dollars (numerical): the price the work of art sold for, in dollars
Included because (string): the reason the work of art was included in the exhibition