Add files via upload
US 2016 Presidential Election Results by State
Description: data about the 2016 U.S. Presidential election results, by state.

Columns:

State or district (string): name of the state or district with votes cast for candidates
Number of votes for Hillary Clinton (numerical): how many votes in total were cast in that state or district for the Democratic candidate, Hillary Clinton
Percentage of votes for Hillary Clinton (numerical): the percentage of votes cast for Hillary Clinton compared to overall votes cast in that state or district
Electoral college votes won for Hillary Clinton (numerical): how many electoral college votes were won in that state or district by Hillary Clinton
Number of votes for Donald Trump (numerical): how many votes in total were cast in that state or district for the Republican candidate, Donald Trump
Percentage of votes for Donald Trump (numerical): the percentage of votes cast for Donald Trump compared to overall votes cast in that state or district
Electoral college votes won for Donald Trump (numerical): how many electoral college votes were won in that state or district by Donald Trump
Number of votes for Gary Johnson (numerical): how many votes in total were cast in that state or district for the Libertarian candidate, Gary Johnson
Percentage of votes for Gary Johnson (numerical): the percentage of votes cast for Gary Johnson compared to overall votes cast in that state or district
Number of votes for Jill Stein (numerical): how many votes in total were cast in that state or district for the Green Party candidate, Jill Stein
Percentage of votes for Jill Stein (numerical): the percentage of votes cast for Jill Stein compared to overall votes cast in that state or district
Margin of votes (numerical): the difference in votes in that state between the winner of the election, Donald Trump, compared to Hillary Clinton. A positive margin indicates a Trump victory, while a negative margin indicates a Clinton victory.
Percentage margin (numerical): the difference in votes in a percentage form in that state, between the winner of the election, Donald Trump, compared to Hillary Clinton. A positive margin indicates a Trump victory, while a negative margin indicates a Clinton victory.
Total (numerical): total number of votes cast in that state or district