Add files via upload
What is this?: This file contains all of the measurements for 80 pairs of blue jeans from the most popular and widely available brands in the US.
Source(s): All data was collected from manual measurements by Jan Diehm and Amber Thomas at brick and mortar stores in Nashville, New York, and Seattle. All measurements of front pockets were taken of the right-hand-side front pocket of empty jeans.
Last Modified: August 13, 2018
Contact Information: Amber Thomas
Spatial Applicability: United States
Temporal Applicability: All measurements were collected between June 29 and August 6, 2018
Observations (Rows): Each row represents data from a single pair of jeans.
Variables (Columns):
Header	Description	Data Type
brand	The full brand name.	text
style	The cut of each pair of jeans (in our analysis, we combined straight and boot-cut styles and skinny and slim styles, but these remain separated here).	text
menWomen	Whether the jeans were listed as "men's" or "women's".	text
name	The name of the specific style of measured pair of jeans as indicated by the tag. (e.g., Fave Super Skinny Jean).	text
brandSize	The size of jeans we measured. Each size listed here reflects the sizing for each brand that is closest to a 32 inch waistband (as indicated by the brand's website).	text
waistSize	The waistband size (in inches) of each measured pair as reported on the brand's website.	number
fabric	The comma-separated fabric combination as indicated by the tag on the jeans. (e.g., 78% cotton, 20% polyester, 2% spandex).	text
price	The price (in US dollars) of each pair of jeans. This number is listed without the dollar sign ($) or any extra 0's (so $39.50 would be written as 39.5)	number
maxHeightFront	The height (in cm) of the front pocket from the top of the highest rivet to the lowest point of the pocket (along the left-hand side or zipper side). See diagram below.	number
minHeightFront	The height (in cm) of the front pocket from the top of the highest rivet to the lowest point of the pocket (along the right-hand side or non-zipper side). See diagram below.	number
rivetHeightFront	The distance (in cm) between the top of the highest rivet to the lowest rivet.	number
maxWidthFront	The width (in cm) from the widest point of the front pocket. See diagram below.	number
minWidthFront	The width (in cm) from the highest rivet to the right or non-zipper side of the pocket.	number
maxHeightBack	The height (in cm) from the deepest point of the back pocket (usually in the pocket's center) to the top of the pocket. See diagram below.	number
minHeightBack	The height (in cm) from the shallowest point of the back pocket to the top of the pocket.	number
maxWidthBack	The width of the pocket at the very top (the opening) See diagram below.	number
minWidthBack	The width of the pocket at its narrowest (just before the pocket tapers to a point). See diagram below.	number
Diagram of Measurements:
https://github.com/the-pudding/data/raw/master/pockets/diagram.png